requirements.txt
decorator==4.0.11
coreapi==2.3.3
Django==1.11.6
DjangoRestFramework==3.7.1
django-allowedsites==0.1.0
django-webtest==1.9.2
ipython==5.1.0
pytest==3.1.3
psycopg2==2.6.1
uWSGI==2.0.14
supervisor==3.2.0